refactor: extract GitHub API queries
