+++ src/App.jsx
@@ -5,6 +5,7 @@ import { Card, CardHeader, CardContent, CardTitle } from "./components/ui/card";
 import { Button } from "./components/ui/button";
 import { Input } from "./components/ui/input";
 import { Loader2, Github, RefreshCcw, ShieldQuestion, Sun, Moon, Monitor, Maximize2, Minimize2, ExternalLink } from "lucide-react";
+import { githubGraphQL, ORG_REPOS_ISSUES, fetchProjectsWithStatus, fetchIssueTypes } from "./api/github";
 
 // Import components
 const Navigation = lazy(() => import("./components/Navigation"));
@@ -30,204 +31,6 @@ const AllIssues = lazy(() => import("./components/AllIssues"));
  * - This is a front-end only app. The token is stored only in memory; nothing is sent anywhere except to api.github.com directly from the browser.
  * - For a production app, add pagination, caching, error handling, and stricter rate-limit management.
  */
-
-const GQL_ENDPOINT = "https://api.github.com/graphql";
-
-// --- Helpers ---------------------------------------------------------------
-async function githubGraphQL(token, query, variables = {}) {
-  const res = await fetch(GQL_ENDPOINT, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({ query, variables }),
-  });
-  if (!res.ok) throw new Error(`GitHub GraphQL error: ${res.status}`);
-  const data = await res.json();
-  if (data.errors) {
-    const msg = data.errors.map((e) => e.message).join("; ");
-    throw new Error(msg || "Unknown GraphQL error");
-  }
-  return data.data;
-}
-
-// --- GraphQL Queries -------------------------------------------------------
-const ORG_REPOS_ISSUES = `
-  query OrgReposIssues($org: String!, $after: String) {
-    organization(login: $org) {
-      name
-      url
-      repositories(first: 30, after: $after, orderBy: {field: PUSHED_AT, direction: DESC}) {
-        pageInfo { hasNextPage endCursor }
-        nodes {
-          id
-          name
-          nameWithOwner
-          url
-          issues(first: 100, orderBy: {field: UPDATED_AT, direction: DESC}, states: [OPEN, CLOSED]) {
-            nodes {
-              id
-              number
-              title
-              body
-              url
-              state
-              createdAt
-              closedAt
-              repository { nameWithOwner url }
-              assignees(first: 10) { nodes { login avatarUrl url } }
-              labels(first: 20) { nodes { id name color } }
-              milestone { id title url dueOn description }
-              issueType { id name color }
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
-const ORG_PROJECTS = `
-  query OrgProjects($org: String!, $after: String) {
-    organization(login: $org) {
-      projectsV2(first: 20, after: $after) {
-        nodes { id number title url }
-        pageInfo { hasNextPage endCursor }
-      }
-    }
-  }
-`;
-
-const PROJECT_ITEMS = `
-  query ProjectItems($pid: ID!, $after: String) {
-    node(id: $pid) {
-      ... on ProjectV2 {
-        number
-        title
-        url
-        items(first: 100, after: $after) {
-          nodes {
-            content {
-              __typename
-              ... on Issue {
-                id
-                number
-                title
-                url
-                state
-                createdAt
-                repository { nameWithOwner }
-                issueType { id name color }
-              }
-            }
-            id
-            fieldValues(first: 20) {
-              nodes {
-                __typename
-                ... on ProjectV2ItemFieldSingleSelectValue {
-                  name
-                  field {
-                    __typename
-                    ... on ProjectV2SingleSelectField { name }
-                  }
-                }
-              }
-            }
-          }
-          pageInfo { hasNextPage endCursor }
-        }
-        fields(first: 20) {
-          nodes {
-            ... on ProjectV2SingleSelectField {
-              id
-              name
-              options {
-                id
-                name
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
-async function fetchProjectsWithStatus(token, org) {
-  let projects = [];
-  let cursor = null;
-  while (true) {
-    const data = await githubGraphQL(token, ORG_PROJECTS, { org, after: cursor });
-    const nodes = data.organization?.projectsV2?.nodes || [];
-    projects = projects.concat(nodes);
-    if (!data.organization?.projectsV2?.pageInfo?.hasNextPage) break;
-    cursor = data.organization.projectsV2.pageInfo.endCursor;
-  }
-
-  const out = [];
-  for (const proj of projects) {
-    let items = [];
-    let after = null;
-    let statusFieldId = null;
-    let statusOptions = {};
-    while (true) {
-      const data = await githubGraphQL(token, PROJECT_ITEMS, { pid: proj.id, after });
-      const node = data.node;
-      if (!statusFieldId) {
-        const fields = node?.fields?.nodes || [];
-        const statusField = fields.find(f => (f.name || '').toLowerCase() === 'status');
-        if (statusField) {
-          statusFieldId = statusField.id;
-          statusOptions = Object.fromEntries((statusField.options || []).map(o => [o.name, o.id]));
-        }
-      }
-      const nodes = node?.items?.nodes || [];
-      const rows = nodes
-        .map((item) => {
-          const issue = item.content;
-          if (!issue || issue.__typename !== "Issue") return null;
-          const status = (item.fieldValues?.nodes || [])
-            .filter((f) => f.__typename === "ProjectV2ItemFieldSingleSelectValue")
-            .map((f) => {
-              const fname = (f.field?.name || "").toLowerCase();
-              return fname === "status" ? f.name : null;
-            })
-            .find(Boolean) || null;
-          return {
-            id: issue.id,
-            number: issue.number,
-            title: issue.title,
-            url: issue.url,
-            state: issue.state,
-            createdAt: issue.createdAt,
-            repository: issue.repository?.nameWithOwner || "",
-            project_status: status,
-            issueType: issue.issueType || null,
-            project_item_id: item.id,
-          };
-        })
-        .filter(Boolean);
-      items = items.concat(rows);
-      if (!node?.items?.pageInfo?.hasNextPage) break;
-      after = node.items.pageInfo.endCursor;
-    }
-    out.push({ id: proj.id, number: proj.number, title: proj.title, url: proj.url, issues: items, statusFieldId, statusOptions });
-  }
-  return out;
-}
-
-async function fetchIssueTypes(token, org) {
-  const res = await fetch(`https://api.github.com/orgs/${encodeURIComponent(org)}/issue-types`, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-      Accept: "application/vnd.github+json",
-    },
-  });
-  if (!res.ok) throw new Error(`GitHub REST error: ${res.status}`);
-  const data = await res.json();
-  return data?.issue_types || data;
-}
 
 // --- Main App -------------------------------------------------------------
 export default function App() {
